Update to Prolog Cafe 1.4
Prolog Cafe is now built with Buck and has a slightly smaller
runtime. Many features not required for Gerrit Code Review were
stripped out of the fork.

Package names were reorganized to make the runtime smaller to
read by pulling exceptions out to their own package. This is
a breaking change for any plugins that contribute a predicate.

+++ gerrit-server/src/main/java/com/google/gerrit/server/change/TestSubmitRule.java
@@ -106,7 +106,7 @@ public class TestSubmitRule implements RestModifyView<RevisionResource, Input> {
     Map<String, None> need;
     Map<String, AccountInfo> may;
     Map<String, None> impossible;
-    Integer prologReductionCount;
+    Long prologReductionCount;
 
     Record(SubmitRecord r, AccountLoader accounts) {
       this.status = r.status;
